Lint

+++ rust/abacus-core/src/db/abacus_db.rs
@@ -403,7 +403,7 @@ impl AbacusDB {
     /// Mark leaf as processed
     pub fn mark_leaf_as_processed(&self, leaf_index: u32) -> Result<(), DbError> {
         debug!(leaf_index = ?leaf_index, "mark leaf as processed");
-        self.store_keyed_encodable(LEAF_PROCESS_STATUS, &leaf_index, &(1 as u32))
+        self.store_keyed_encodable(LEAF_PROCESS_STATUS, &leaf_index, &(1_u32))
     }
 
     /// Retrieve leaf processing status
